Fully attach or load packages for making model predictions

Some models require one or more R packages to be fully attached to make predictions, and some require only that the namespace of one or more R packages is loaded.

attach_pkgs(pkgs)

load_pkgs(pkgs)

Arguments

pkgs
A character vector of package names to load or fully attach.

Value

An invisible TRUE.

Details

These two functions will attempt either to:

fully attach or

load

the namespace of the pkgs vector of package names, preserving the current random seed.

To learn more about load vs. attach, read the NAMESPACE chapter of R Packages. For deploying a model, it is likely safer to fully attach needed packages but that comes with the risk of naming conflicts between packages.

Examples

if (FALSE) {
## succeeds
attach_pkgs(c("knitr", "readr"))
## fails
attach_pkgs(c("bloopy", "readr"))
## succeeds
load_pkgs(c("knitr", "readr"))
## fails
load_pkgs(c("bloopy", "readr"))
}