Model handler functions for API endpoint

These are developer-facing functions, useful for supporting new model types. Each model supported by vetiver_model() uses two handler functions in vetiver_pr_predict():

The handler_startup function executes when the API starts. Use this function for tasks like loading packages. A model can use the default method here, which is NULL (to do nothing at startup).

The handler_predict function executes at each API call. Use this function for calling predict() and any other tasks that must be executed at each API call.

handler_startup(vetiver_model)

# S3 method for default
handler_startup(vetiver_model)

handler_predict(vetiver_model, ...)

# S3 method for default
handler_predict(vetiver_model, ...)

# S3 method for lm
handler_predict(vetiver_model, ...)

# S3 method for xgb.Booster
handler_startup(vetiver_model)

# S3 method for xgb.Booster
handler_predict(vetiver_model, ...)

Arguments

vetiver_model
A deployable vetiver_model() object

...
Other arguments passed to predict(), such as prediction type

Value

A handler_startup function should return invisibly, while a handler_predict function should return a function with the signature function(req).

Details

These are two generics that use the class of vetiver_model$model for dispatch.

Examples

cars_lm <- lm(mpg ~ ., data = mtcars)
v <- vetiver_model(cars_lm, "cars_linear")
handler_startup(v)
handler_predict(v)

#> function (req) 
#> {
#>     newdata <- req$body
#>     if (!is_null(ptype)) {
#>         newdata <- vetiver_type_convert(newdata, ptype)
#>         newdata <- hardhat::scream(newdata, ptype)
#>     }
#>     ret <- predict(vetiver_model$model, newdata = newdata, ...)
#>     list(.pred = ret)
#> }
#> <bytecode: 0x55a2b6084ba8>
#> <environment: 0x55a2b6085c80>