Read and write a trained model to a board of models

Use vetiver_pin_write() to pin a trained model to a board of models, along with an input prototype for new data and other model metadata. Use vetiver_pin_read() to retried that pinned object.

vetiver_pin_write(vetiver_model)

vetiver_pin_read(board, name, version = NULL)

Arguments

vetiver_model
A deployable vetiver_model() object

board
A pin board, created by board_folder(), board_rsconnect(), board_url() or other other board_ function.

name
Pin name.

version
Retrieve a specific version of a pin. Use pin_versions() to find out which versions are available and when they were created.

Value

vetiver_pin_read() returns a vetiver_model(); vetiver_pin_write() returns the name of the new pin, invisibly.

Details

These functions read and write a vetiver_model() pin on the specified board containing the model object itself and other elements needed for prediction, such as the model's input data prototype or which packages are needed at prediction time. You may use pins::pin_read() or pins::pin_meta() to handle the pin, but vetiver_pin_read() returns a vetiver_model() object ready for deployment.

Examples

library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
v <- vetiver_model(cars_lm, "cars_linear", model_board)
vetiver_pin_write(v)

#> Creating new version '20211008T224440Z-92526'

#> Writing to pin 'cars_linear'

model_board

#> Pin board <pins_board_folder>
#> Path: '/tmp/RtmphVwLm7/pins-431a1587b80f'
#> Cache size: 0
#> Pins [1]: 'cars_linear'

vetiver_pin_read(model_board, "cars_linear")

#> 
#> ── cars_linear ─ <butchered_lm> model for deployment 
#> An OLS linear regression model using 10 features

# can use `version` argument to read a specific version:
pin_versions(model_board, "cars_linear")

#> # A tibble: 1 × 3
#>   version                created             hash 
#>   <chr>                  <dttm>              <chr>
#> 1 20211008T224440Z-92526 2021-10-08 22:44:00 92526