Create a Plumber API to predict with a deployable vetiver_model() object

Use vetiver_pr_predict() to add a POST endpoint for predictions from a trained, pinned vetiver_model() to a Plumber router.

vetiver_pr_predict(
  pr,
  vetiver_model,
  path = "/predict",
  debug = is_interactive(),
  ...
)

Arguments

pr
A Plumber router, such as from plumber::pr().

vetiver_model
A deployable vetiver_model() object

path
The endpoint path

debug
TRUE provides more insight into your API errors.

...
Other arguments passed to predict(), such as prediction type

Value

A Plumber router with the prediction endpoint added.

Details

First store and version your vetiver_model() with vetiver_pin_write(), and then create an API endpoint with vetiver_pr_predict().

Setting debug = TRUE may expose any sensitive data from your model in API errors.

Examples

library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
v <- vetiver_model(cars_lm, "cars_linear", model_board)
vetiver_pin_write(v)

#> Creating new version '20211019T154516Z-92526'

#> Writing to pin 'cars_linear'

library(plumber)
pr() %>% vetiver_pr_predict(v)

#> # Plumber router with 1 endpoint, 4 filters, and 0 sub-routers.
#> # Use `pr_run()` on this object to start the API.
#> ├──[queryString]
#> ├──[body]
#> ├──[cookieParser]
#> ├──[sharedSecret]
#> └──/predict (POST)
#> 

## next, pipe to `pr_run()`