Write a deployable Plumber file for a vetiver model

Use vetiver_write_plumber() to create a plumber.R file for a vetiver_model() that has been versioned and stored via vetiver_pin_write().

vetiver_write_plumber(
  board,
  name,
  version = NULL,
  ...,
  file = "plumber.R",
  docs = "rapidoc"
)

Arguments

board
A pin board, created by board_folder(), board_rsconnect(), board_url() or other other board_ function.

name
Pin name.

version
Retrieve a specific version of a pin. Use pin_versions() to find out which versions are available and when they were created.

...
Other arguments passed to vetiver_pr_predict() such as the endpoint path or prediction type.

file
A path to write the Plumber file. Defaults to plumber.R in the working directory. See plumber::plumb() for naming precedence rules.

docs
A character value matching a package for OpenAPI visual documentation such as rapidoc (the default) or redoc. Use NULL for default Plumber docs.

Value

The content of the plumber.R file, invisibly.

Details

By default, this function will find and use the latest version of vetiver model; the model API (when deployed) will be linked to that specific version. You can override this default behavior by choosing a specific version.

Examples

library(pins)
tmp <- tempfile()
b <- board_temp(versioned = TRUE)
cars_lm <- lm(mpg ~ ., data = mtcars)
v <- vetiver_model(cars_lm, "cars_linear", b)
vetiver_pin_write(v)

#> Creating new version '20211008T222613Z-92526'

#> Writing to pin 'cars_linear'

vetiver_write_plumber(b, "cars_linear", file = tmp)