Metadata constructors for vetiver_model() object

These are developer-facing functions, useful for supporting new model types. The metadata stored in a vetiver_model() object has four elements:

$user, the metadata supplied by the user

$version, the version of the pin (which can be NULL before pinning)

$url, the URL where the pin is located, if any

$required_pkgs, a character string of R packages required for prediction

vetiver_meta(user = list(), version = NULL, url = NULL, required_pkgs = NULL)

vetiver_create_meta(model, metadata)

# S3 method for default
vetiver_create_meta(model, metadata)

# S3 method for xgb.Booster
vetiver_create_meta(model, metadata)

Arguments

user
Metadata supplied by the user

version
Version of the pin

url
URL for the pin, if any

required_pkgs
Character string of R packages required for prediction

model
A trained model, such as an lm() model or a tidymodels workflows::workflow().

metadata
A list containing additional metadata to store with the pin. When retrieving the pin, this will be stored in the user key, to avoid potential clashes with the metadata that pins itself uses.

Value

The vetiver_meta() constructor returns a list. The vetiver_create_meta function returns a vetiver_meta() list.

Examples

vetiver_meta()
#> $user
#> list()
#> 
#> $version
#> NULL
#> 
#> $url
#> NULL
#> 
#> $required_pkgs
#> NULL
#> 

cars_lm <- lm(mpg ~ ., data = mtcars)
vetiver_create_meta(cars_lm, list())
#> $user
#> list()
#> 
#> $version
#> NULL
#> 
#> $url
#> NULL
#> 
#> $required_pkgs
#> NULL
#>